+++ src/main/java/org/truffleruby/core/numeric/IntegerNodes.java
@@ -583,14 +583,12 @@ public abstract class IntegerNodes {
         Object mod(long a, RubyBignum b) {
             // TODO(CS): why are we getting this case?
 
-            long mod = BigInteger.valueOf(a).mod(b.value).longValue();
-
-            if (mod < 0 && b.value.compareTo(BigInteger.ZERO) > 0 ||
-                    mod > 0 && b.value.compareTo(BigInteger.ZERO) < 0) {
-                return createBignum(BigInteger.valueOf(mod).add(b.value));
+            if ((a < 0 && b.value.compareTo(BigInteger.ZERO) > 0) ||
+                    (a >= 0 && b.value.compareTo(BigInteger.ZERO) < 0)) {
+                return createBignum(BigInteger.valueOf(a).add(b.value));
             }
 
-            return mod;
+            return a;
         }
 
         @TruffleBoundary